DB 연결 및 디젤 설정 완료

+++ src/schema.rs
@@ -1,6 +1,6 @@
 table! {
     test (id) {
-        id -> Integer, 
+        id -> BigInt, 
         text -> Text,
     }
 }
